oci-notification-telegram-forwarder/requirements.txt
fdk>=0.1.102
python-telegram-bot